feat(validations) project validations  - add validations to the controllers
[Finishes #165303327]
