+++ api-routes/index.js
@@ -4,7 +4,10 @@ const router = require('express').Router();
 // import signup
 const signupController = require('../controllers/auth/signup');
 const loginController = require('../controllers/auth/login');
-const bankController = require('../controllers/bank-account');
+const bankTransactions = require('../controllers/bank-account/transactions');
+const deleteAccount = require('../controllers/bank-account/deleteAccount');
+const accountStatus = require('../controllers/bank-account/accountStatus');
+const createBankAccount = require('../controllers/bank-account/createAccount');
 
 // import VerifyToken middleware function
 const middleware = require('../middleware');
@@ -15,14 +18,14 @@ router.route('/signup')
 router.route('/login')
   .post(loginController.login);
 router.route('/accounts')
-  .post(middleware.verifyToken, bankController.createBankAccount);
+  .post(middleware.verifyToken, createBankAccount);
 router.route('/account/:accountNumber')
-  .patch(middleware.verifyToken, bankController.accountStatus)
-  .delete(middleware.verifyToken, bankController.deleteAccount);
+  .patch(middleware.verifyToken, accountStatus)
+  .delete(middleware.verifyToken, deleteAccount);
 router.route('/transactions/:accountNumber/credit')
-  .post(middleware.verifyToken, bankController.creditTransaction);
+  .post(middleware.verifyToken, bankTransactions.creditTransaction);
 router.route('/transactions/:accountNumber/debit')
-  .post(middleware.verifyToken, bankController.debitTransaction);
+  .post(middleware.verifyToken, bankTransactions.debitTransaction);
 router.route('/users')
   .get(middleware.verifyToken, loginController.allUsers);
 // Export API routes
