+++ controllers/auth/login.js
@@ -1,3 +1,4 @@
+/* eslint-disable consistent-return */
 const bcrypt = require('bcryptjs');
 // import token generator
 const middleware = require('../../middleware');
@@ -7,13 +8,21 @@ const { users } = require('../../models');
 // user login
 exports.login = (req, res) => {
   const { password, email } = req.body;
+  // Email and Password are required
+  if (!email || !password) {
+    return res.status(400).json({
+      status: 400,
+      error: 'Email, Password and type are required !',
+    });
+  }
+
   const isAuthenticated = users.some(
     user => bcrypt.compareSync(password, user.password) && user.email === email,
   );
 
   if (isAuthenticated !== true) {
     // Wrong password
-    res.json({
+    return res.json({
       status: 401,
       error: 'Wrong email or password',
     });
@@ -21,10 +30,8 @@ exports.login = (req, res) => {
 
   users.forEach((user) => {
     if (user.email === email) {
-    //
-
       // return the JWT token for the future API calls
-      res.json({
+      return res.json({
         status: 200,
         data: {
           token: middleware.token(user.id),
@@ -39,10 +46,8 @@ exports.login = (req, res) => {
 };
 
 // fetch all users
-exports.allUsers = (req, res) => {
-  res.status(200).json({
-    status: 200,
-    data: users,
-    id: req.userId,
-  });
-};
+exports.allUsers = (req, res) => res.status(200).json({
+  status: 200,
+  data: users,
+  id: req.userId,
+});
